// dashboard/src/api/models/createUserRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */

export interface CreateUserRequest {
  username: string;
  displayName: string;
  password: string;
  /** Optional initial roles to assign to the user. */
  roles?: string[];
}